+++ pyccel/parser/scope.py
@@ -12,6 +12,7 @@ from pyccel.ast.datatypes import InhomogeneousTupleType
 from pyccel.ast.headers   import MacroFunction, MacroVariable
 from pyccel.ast.headers   import FunctionHeader, MethodHeader
 from pyccel.ast.internals import PyccelSymbol, PyccelFunction
+from pyccel.ast.typingext import TypingTypeVar
 from pyccel.ast.variable  import Variable, DottedName, AnnotatedPyccelSymbol
 from pyccel.ast.variable  import IndexedElement, DottedVariable
 
@@ -69,7 +70,7 @@ class Scope(object):
 
     categories = ('functions','variables','classes',
             'imports','symbolic_functions', 'symbolic_aliases',
-            'macros','templates','headers','decorators',
+            'macros','headers','decorators',
             'cls_constructs')
 
     def __init__(self, *, name=None, decorators = (), is_loop = False,
@@ -105,14 +106,6 @@ class Scope(object):
         self._loops = []
 
         self._dotted_symbols = []
-
-    def __setstate__(self, state):
-        state = state[1] # Retrieve __dict__ ignoring None
-        if any(s not in state for s in self.__slots__):
-            raise AttributeError("Missing attribute from slots. Please update pickle file")
-
-        for s in state:
-            setattr(self, s, state[s])
 
     def new_child_scope(self, name, **kwargs):
         """
@@ -188,12 +181,6 @@ class Scope(object):
         """A dictionary of user defined headers which may
         be applied to functions in this scope"""
         return self._locals['headers']
-
-    @property
-    def templates(self):
-        """A dictionary of user defined templates which may
-        be applied to functions in this scope"""
-        return self._locals['templates']
 
     @property
     def decorators(self):
@@ -468,10 +455,6 @@ class Scope(object):
             name = name.name[-1]
 
         self._locals['macros'][name] = macro
-
-    def insert_template(self, expr):
-        """append the scope's templates with the given template"""
-        self._locals['templates'][expr.name] = expr
 
     def insert_header(self, expr):
         """
@@ -780,6 +763,26 @@ class Scope(object):
         imports = list(self._imports['imports'].keys())
         imports.extend([i for s in self._sons_scopes.values() for i in s.collect_all_imports()])
         return imports
+
+    def collect_all_type_vars(self):
+        """
+        Collect all TypeVar objects which are available in this scope.
+
+        Collect all TypeVar objects which are available in this scope. This includes
+        TypeVars declared in parent scopes.
+
+        Returns
+        -------
+        list[TypeVar]
+            A list of TypeVars in the scope.
+        """
+        type_vars = {n:t for n,t in self.symbolic_aliases.items() if isinstance(t, TypingTypeVar)}
+        if self.parent_scope:
+            parent_type_vars = self.parent_scope.collect_all_type_vars()
+            parent_type_vars.update(type_vars)
+            return parent_type_vars
+        else:
+            return type_vars
 
     def update_parent_scope(self, new_parent, is_loop, name = None):
         """ Change the parent scope