+++ tests/semantic/scripts/imports.py
@@ -1,9 +1,0 @@
-# pylint: disable=missing-function-docstring, missing-module-docstring
-from expressions import ai
-xi = ai + 1
-
-from expressions import ad, bd
-xd = ad * bd
-
-from functions import decr
-yi = decr(xi)